+++ internal/endtoend/testdata/typed_dml/expected/db/queries.sql.go
@@ -100,7 +100,7 @@ func bindUpdateRecordsRowsItem(values []UpdateRecordsRowsItem) types.Value {
 func (q *Queries) ListRecords(ctx context.Context, arg uint64) ([]ListRecordsRow, error) {
 	rows, err := q.db.QueryContext(ctx, ""+
 		"DECLARE $owner_id AS Uint64;\n"+
-		"SELECT owner_hash, owner_id, record_id, group_id, payload, attributes, created_at, updated_at\n"+
+		"SELECT owner_hash, owner_id, record_id, group_id, payload, attributes, created_at, updated_at, note\n"+
 		"FROM records\n"+
 		"WHERE owner_hash = Digest::CityHash(CAST($owner_id AS String))\n"+
 		"ORDER BY record_id;",
@@ -123,6 +123,7 @@ func (q *Queries) ListRecords(ctx context.Context, arg uint64) ([]ListRecordsRow
 			&row.Attributes,
 			&row.CreatedAt,
 			&row.UpdatedAt,
+			&row.Note,
 		); err != nil {
 			return nil, err
 		}
@@ -252,4 +253,153 @@ func bindGetRecordKey(item GetRecordKey) types.Value {
 		types.StructFieldValue("owner_hash", types.Uint64Value(item.OwnerHash)),
 		types.StructFieldValue("record_id", types.TextValue(item.RecordID)),
 	)
+}
+
+// -- name: InsertNamedRecords :exec
+func (q *Queries) InsertNamedRecords(ctx context.Context, arg InsertNamedRecordsParams) error {
+	_, err := q.db.ExecContext(ctx, ""+
+		"DECLARE $owner_id AS Uint64;\n"+
+		"DECLARE $rows AS List<Struct<payload: Bytes, attributes: Json, group_id: Utf8, record_id: Utf8, note: Utf8,>>;\n"+
+		"DECLARE $created_at AS Timestamp;\n"+
+		"INSERT INTO records\n"+
+		"SELECT\n"+
+		"    $created_at AS updated_at,\n"+
+		"    r.`payload` AS `payload`, `r`.`attributes` AS `attributes`, `r`.`group_id` AS `group_id`, `r`.`record_id` AS `record_id`, `r`.`note` AS `note`,\n"+
+		"    Digest::CityHash(CAST($owner_id AS String)) AS owner_hash,\n"+
+		"    $owner_id AS owner_id,\n"+
+		"    $created_at AS created_at\n"+
+		"FROM AS_TABLE($rows) AS r;",
+		sql.Named("owner_id", arg.OwnerID),
+		sql.Named("rows", bindInsertNamedRecordsRowsItem(arg.Rows)),
+		sql.Named("created_at", arg.CreatedAt),
+	)
+
+	return err
+}
+
+func bindInsertNamedRecordsRowsItem(values []InsertNamedRecordsRowsItem) types.Value {
+	if len(values) == 0 {
+		return types.ZeroValue(types.List(types.Struct(
+			types.StructField("payload", types.TypeBytes),
+			types.StructField("attributes", types.TypeJSON),
+			types.StructField("group_id", types.TypeText),
+			types.StructField("record_id", types.TypeText),
+			types.StructField("note", types.TypeText),
+		)))
+	}
+	items := make([]types.Value, len(values))
+	for i, item := range values {
+		items[i] = types.StructValue(
+			types.StructFieldValue("payload", types.BytesValue(item.Payload)),
+			types.StructFieldValue("attributes", types.JSONValue(item.Attributes)),
+			types.StructFieldValue("group_id", types.TextValue(item.GroupID)),
+			types.StructFieldValue("record_id", types.TextValue(item.RecordID)),
+			types.StructFieldValue("note", types.TextValue(item.Note)),
+		)
+	}
+	return types.ListValue(items...)
+}
+
+// -- name: UpsertNamedRecords :exec
+func (q *Queries) UpsertNamedRecords(ctx context.Context, arg UpsertNamedRecordsParams) error {
+	_, err := q.db.ExecContext(ctx, ""+
+		"DECLARE $owner_hash AS Uint64;\n"+
+		"DECLARE $owner_id AS Uint64;\n"+
+		"DECLARE $created_at AS Timestamp;\n"+
+		"DECLARE $rows AS List<Struct<body: Bytes, key: Utf8, group_id: Utf8, attributes: Json,>>;\n"+
+		"UPSERT INTO records\n"+
+		"SELECT r.body AS payload, r.key AS record_id, $owner_hash AS owner_hash,\n"+
+		"    $owner_id AS owner_id, r.group_id, r.attributes,\n"+
+		"    $created_at AS created_at, $created_at AS updated_at\n"+
+		"FROM AS_TABLE($rows) AS r;",
+		sql.Named("owner_hash", arg.OwnerHash),
+		sql.Named("owner_id", arg.OwnerID),
+		sql.Named("created_at", arg.CreatedAt),
+		sql.Named("rows", bindUpsertNamedRecordsRowsItem(arg.Rows)),
+	)
+
+	return err
+}
+
+func bindUpsertNamedRecordsRowsItem(values []UpsertNamedRecordsRowsItem) types.Value {
+	if len(values) == 0 {
+		return types.ZeroValue(types.List(types.Struct(
+			types.StructField("body", types.TypeBytes),
+			types.StructField("key", types.TypeText),
+			types.StructField("group_id", types.TypeText),
+			types.StructField("attributes", types.TypeJSON),
+		)))
+	}
+	items := make([]types.Value, len(values))
+	for i, item := range values {
+		items[i] = types.StructValue(
+			types.StructFieldValue("body", types.BytesValue(item.Body)),
+			types.StructFieldValue("key", types.TextValue(item.Key)),
+			types.StructFieldValue("group_id", types.TextValue(item.GroupID)),
+			types.StructFieldValue("attributes", types.JSONValue(item.Attributes)),
+		)
+	}
+	return types.ListValue(items...)
+}
+
+// -- name: UpsertWildcardRecords :exec
+func (q *Queries) UpsertWildcardRecords(ctx context.Context, arg []UpsertWildcardRecordsRowsItem) error {
+	_, err := q.db.ExecContext(ctx, ""+
+		"DECLARE $rows AS List<Struct<payload: Bytes, record_id: Utf8, owner_hash: Uint64, group_id: Utf8, owner_id: Uint64, attributes: Json, created_at: Timestamp, updated_at: Timestamp,>>;\n"+
+		"UPSERT INTO records\n"+
+		"SELECT `payload`, `record_id`, `owner_hash`, `group_id`, `owner_id`, `attributes`, `created_at`, `updated_at` FROM AS_TABLE($rows);",
+		sql.Named("rows", bindUpsertWildcardRecordsRowsItem(arg)),
+	)
+
+	return err
+}
+
+func bindUpsertWildcardRecordsRowsItem(values []UpsertWildcardRecordsRowsItem) types.Value {
+	if len(values) == 0 {
+		return types.ZeroValue(types.List(types.Struct(
+			types.StructField("payload", types.TypeBytes),
+			types.StructField("record_id", types.TypeText),
+			types.StructField("owner_hash", types.TypeUint64),
+			types.StructField("group_id", types.TypeText),
+			types.StructField("owner_id", types.TypeUint64),
+			types.StructField("attributes", types.TypeJSON),
+			types.StructField("created_at", types.TypeTimestamp),
+			types.StructField("updated_at", types.TypeTimestamp),
+		)))
+	}
+	items := make([]types.Value, len(values))
+	for i, item := range values {
+		items[i] = types.StructValue(
+			types.StructFieldValue("payload", types.BytesValue(item.Payload)),
+			types.StructFieldValue("record_id", types.TextValue(item.RecordID)),
+			types.StructFieldValue("owner_hash", types.Uint64Value(item.OwnerHash)),
+			types.StructFieldValue("group_id", types.TextValue(item.GroupID)),
+			types.StructFieldValue("owner_id", types.Uint64Value(item.OwnerID)),
+			types.StructFieldValue("attributes", types.JSONValue(item.Attributes)),
+			types.StructFieldValue("created_at", types.TimestampValueFromTime(item.CreatedAt)),
+			types.StructFieldValue("updated_at", types.TimestampValueFromTime(item.UpdatedAt)),
+		)
+	}
+	return types.ListValue(items...)
+}
+
+// -- name: ClearNamedRecordNote :one
+func (q *Queries) ClearNamedRecordNote(ctx context.Context, arg ClearNamedRecordNoteParams) (ClearNamedRecordNoteRow, error) {
+	var row ClearNamedRecordNoteRow
+	err := q.db.QueryRowContext(ctx, ""+
+		"DECLARE $owner_hash AS Uint64;\n"+
+		"DECLARE $record_id AS Utf8;\n"+
+		"UPSERT INTO records\n"+
+		"SELECT record_id, owner_hash, owner_id, group_id, payload, attributes, created_at, updated_at, NULL AS note\n"+
+		"FROM records\n"+
+		"WHERE owner_hash = $owner_hash AND record_id = $record_id\n"+
+		"RETURNING record_id, note;",
+		sql.Named("owner_hash", arg.OwnerHash),
+		sql.Named("record_id", arg.RecordID),
+	).Scan(
+		&row.RecordID,
+		&row.Note,
+	)
+
+	return row, err
 }